+++ app/scripts/components/app/footer.js
@@ -1,18 +1,15 @@
 'use strict';
 
 import React from 'react';
-import createReactClass from 'create-react-class';
 
-var Footer = createReactClass({
-  displayName: 'Footer',
-
-  render: function () {
+class Footer extends React.Component {
+  render () {
     return (
       <div className='footer'>
         <h1>Footer</h1>
       </div>
     );
   }
-});
+}
 
 export default Footer;