+++ app/scripts/components/collections/add.js
@@ -1,7 +1,6 @@
 'use strict';
 import React from 'react';
 import PropTypes from 'prop-types';
-import createReactClass from 'create-react-class';
 import { connect } from 'react-redux';
 import { createCollection } from '../../actions';
 import { getCollectionId, collectionHref } from '../../utils/format';
@@ -15,12 +14,8 @@ const getBaseRoute = function (collectionId) {
   }
 };
 
-var AddCollection = createReactClass({
-  propTypes: {
-    collections: PropTypes.object
-  },
-
-  render: function () {
+class AddCollection extends React.Component {
+  render () {
     return (
       <AddRaw
         pk={'new-collection'}
@@ -33,7 +28,11 @@ var AddCollection = createReactClass({
       />
     );
   }
-});
+}
+
+AddCollection.propTypes = {
+  collections: PropTypes.object
+};
 
 export default connect(state => ({
   collections: state.collections