+++ app/scripts/components/form/batch-async-command.js
@@ -2,38 +2,33 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import queue from 'stubborn-queue';
-import createReactClass from 'create-react-class';
 import AsyncCommand from './async-command';
 import { updateDelay } from '../../config';
 
 const CONCURRENCY = 3;
 const IN_PROGRESS = 'Processing...';
 
-const BatchCommand = createReactClass({
-  propTypes: {
-    action: PropTypes.func,
-    dispatch: PropTypes.func,
-    state: PropTypes.object,
-    text: PropTypes.string,
-    selection: PropTypes.array,
-    className: PropTypes.string,
-    onSuccess: PropTypes.func,
-    onError: PropTypes.func,
-    confirm: PropTypes.func,
-    confirmOptions: PropTypes.array,
-    updateDelay: PropTypes.number
-  },
-
-  getInitialState: function () {
-    return {
+class BatchCommand extends React.Component {
+  constructor () {
+    super();
+    this.state = {
       callbacks: {},
       activeModal: false,
       completed: 0,
       status: null
     };
-  },
+    this.confirm = this.confirm.bind(this);
+    this.cancel = this.cancel.bind(this);
+    this.start = this.start.bind(this);
+    this.initAction = this.initAction.bind(this);
+    this.onComplete = this.onComplete.bind(this);
+    this.createErrorMessage = this.createErrorMessage.bind(this);
+    this.cleanup = this.cleanup.bind(this);
+    this.isInflight = this.isInflight.bind(this);
+    this.handleClick = this.handleClick.bind(this);
+  }
 
-  UNSAFE_componentWillReceiveProps: function (newProps) {
+  UNSAFE_componentWillReceiveProps (newProps) { // eslint-disable-line camelcase
     const { state } = newProps;
     const { callbacks, completed } = this.state;
     // on success or error, call and remove the saved callback
@@ -46,18 +41,18 @@ const BatchCommand = createReactClass({
         this.setState({ callbacks, completed: completed + 1 });
       }
     });
-  },
+  }
 
-  confirm: function () {
+  confirm () {
     if (!this.isInflight()) this.start();
-  },
+  }
 
-  cancel: function () {
+  cancel () {
     // prevent cancel when inflight
     if (!this.isInflight()) this.setState({ activeModal: false });
-  },
+  }
 
-  start: function () {
+  start () {
     const { selection } = this.props;
     // if we have inflight callbacks, don't allow further clicks
     if (!Array.isArray(selection) || !selection.length ||
@@ -67,19 +62,19 @@ const BatchCommand = createReactClass({
       q.add(this.initAction, selection[i]);
     }
     q.done(this.onComplete);
-  },
+  }
 
   // save a reference to the callback in state, then init the action
-  initAction: function (id, callback) {
+  initAction (id, callback) {
     const { dispatch, action } = this.props;
     const { callbacks } = this.state;
     callbacks[id] = callback;
     this.setState({ callbacks });
     return dispatch(action(id));
-  },
+  }
 
   // immediately change the UI to show either success or error
-  onComplete: function (errors, results) {
+  onComplete (errors, results) {
     const delay = this.props.updateDelay ? this.props.updateDelay : updateDelay;
     // turn array of errors from queue into single error for ui
     const error = this.createErrorMessage(errors);
@@ -87,33 +82,33 @@ const BatchCommand = createReactClass({
     setTimeout(() => {
       this.cleanup(error, results);
     }, delay);
-  },
+  }
 
   // combine multiple errors into one
-  createErrorMessage: function (errors) {
+  createErrorMessage (errors) {
     if (!errors || !errors.length) return;
     return `${errors.length} errors occurred: \n${errors.map((err) => err.error.toString()).join('\n')}`;
-  },
+  }
 
   // call onSuccess and onError functions as needed
-  cleanup: function (error, results) {
+  cleanup (error, results) {
     const { onSuccess, onError } = this.props;
     this.setState({ activeModal: false, completed: 0, status: null });
     if (error && typeof onError === 'function') onError(error);
     if (results && results.length && typeof onSuccess === 'function') onSuccess(results);
-  },
+  }
 
-  isInflight: function () {
+  isInflight () {
     return !!Object.keys(this.state.callbacks).length;
-  },
+  }
 
-  handleClick: function () {
+  handleClick () {
     if (this.props.confirm) {
       this.setState({ activeModal: true, completed: 0 });
     } else this.start();
-  },
+  }
 
-  render: function () {
+  render () {
     const { text, selection, className, confirm, confirmOptions } = this.props;
     const { activeModal, completed, status } = this.state;
     const todo = selection.length;
@@ -164,5 +159,20 @@ const BatchCommand = createReactClass({
       </div>
     );
   }
-});
+}
+
+BatchCommand.propTypes = {
+  action: PropTypes.func,
+  dispatch: PropTypes.func,
+  state: PropTypes.object,
+  text: PropTypes.string,
+  selection: PropTypes.array,
+  className: PropTypes.string,
+  onSuccess: PropTypes.func,
+  onError: PropTypes.func,
+  confirm: PropTypes.func,
+  confirmOptions: PropTypes.array,
+  updateDelay: PropTypes.number
+};
+
 export default BatchCommand;