+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Execution details (inputs and outputs) are shown for executions and execution steps. [CUMULUS-692]
+- Link to parent workflow execution if any exists. [Cumulus-689]
 
 ## [v1.2.0] - 2018-08-08