+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Link to parent workflow execution if any exists. [Cumulus-689]
 - Logs for specific workflow executions. [Cumulus-691]
 - Documentation was added to the `execution-status.js` component about how to find task / version information for step function events. [Cumulus-690]
+- Added Cypress for front-end testing. [Cumulus-918]
+- Added tests for the login page and dashboard home page. [Cumulus-638]
 
 ## [v1.2.0] - 2018-08-08