+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added Cypress for front-end testing. [Cumulus-918]
 - Added tests for the login page and dashboard home page. [Cumulus-638]
 - Added tests for the Collections page and Providers page. [Cumulus-643]
+- Added warning message to granules `reingest` button to indicates that existing data will be overwritten. [Cumulus-792]
 
 ## [v1.2.0] - 2018-08-08